Use archived reference for licensing
The library is based on a C implementation of the Madgwick AHRS algorithm provided by Sebastian Madgwick.

Unfortunately the licensing of that base code was not well documented. The licensing was only specified on the website
of Sebastian Madgwick's x-io Technologies website where the code is hosted. Since this important information was
difficult to find, a comment was added to the source code headers which contains a reference link and quote of the
relevant statement.

There were two serious problems with that comment:

* The information was later removed from the linked page.
* The quote did not include the essential link which clarifies exactly which version of the GPL license is offered.

These are resolved by:

* Using a link to an archived version of the reference web page which still provides the licensing information.
* Adding markup to indicate the link in the quoted statement. Since that link relies on a redirect, an archive link is
  also used here in order to avoid future bit rot.