Correct licensing documentation in readme
The project's readme contains a "License" section which summarizes the project licensing.

Previously the information was incorrect (it seems to just be a copy/paste of the content that was present in most of
Arduino's libraries at that time, by someone who didn't actually verify that it was appropriate):

* The copyright assignment for Sebastian Madgwick, the author of the code upon which the library is based was omitted,
  falsely indicating that Arduino LLC is the sole copyright holder.
* It stated that the project is licensed as LGPL 2.1, when it is actually the GPL 3.0 licensing inherited from the base
  code.

The missing copyright assignment is hereby added and the license information corrected.